+++ CHANGELOG.md
@@ -1,5 +1,35 @@
 ## envconsul CHANGELOG
 
+## v.0.11.0 (Nov 30, 2020)
+
+IMPROVEMENTS:
+
+* Add support for Vault Agent token files [[GH-249](https://github.com/hashicorp/envconsul/pull/249)]
+* Update whitelist/blacklist config options to allowlist/denylist with backward compatibility [[GH-246](https://github.com/hashicorp/envconsul/pull/246)]
+* Update Consul Template dependency to v0.25.1 [[GH-245](https://github.com/hashicorp/envconsul/pull/245)]
+
+BUG FIXES:
+
+* Remove code/logic for working with (long deprecated) Vault grace [[GH-245](https://github.com/hashicorp/envconsul/pull/245)]
+
+## v0.10.0 (Aug 10, 2020)
+
+IMPROVEMENTS:
+
+* Allow users to use ENV vars in prefix paths using Go templates [[GH-167](https://github.com/hashicorp/envconsul/pull/167)]
+
+## v0.9.3 (Apr 27, 2020)
+
+BUG FIXES:
+
+* Fix renewing wrapped tokens, renew using unwrapped token [[GH-239](https://github.com/hashicorp/envconsul/pull/239), [GH-222](https://github.com/hashicorp/envconsul/issues/222)]
+
+* Fix issue with secret/prefix entries with no path [[GH-240](https://github.com/hashicorp/envconsul/pull/240), [GH-165](https://github.com/hashicorp/envconsul/issues/165)]
+
+DOCUMENTATION:
+
+* Remove unused -exec-reload-signal from docs [[GH-238](https://github.com/hashicorp/envconsul/pull/238), [GH-237](https://github.com/hashicorp/envconsul/issues/237)]
+
 ## v0.9.2 (Jan 08, 2020)
 
 SECURITY: